requirements.txt
requests
yt-dlp
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow